fix: fixed issue with rheostat.css (#333)
The app was throwing an error when trying to import the SlidingInput component from uibook. This is
because it includes a CSS import statement but uibook is never built with webpack. I spent a whole
day trying to get webpack to build uibook properly but I encountered various issues and eventually
gave up. As a quick fix, we can simply use the next-transpile-modules plugin to do the transpilation
in storefront instead.